requirements.txt
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
billiard==4.2.2
blinker==1.9.0
boto3==1.40.60
botocore==1.40.73
cachetools==6.2.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
convertdate==2.4.0
crispy-bootstrap5==2025.6
cryptography==46.0.3
cssselect==1.3.0
cssselect2==0.8.0
cssutils==2.11.1
diff-match-patch==20241021
Django==5.2.7
django-allauth==65.12.1
django-appconf==1.2.0
django-autocomplete-light==3.12.1
django-ckeditor-5==0.2.18
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-crontab==0.7.1
django-dbbackup==5.0.0
django-debug-toolbar==6.0.0
django-extensions==4.1
django-filter==25.2
django-htmx==1.26.0
django-import-export==4.3.12
django-localflavor==5.0
django-maintenance-mode==0.22.0
django-model-utils==5.0.0
django-phonenumber-field==7.3.0
django-premailer==0.2.0
django-qr-code==4.2.0
django-recaptcha==4.1.0
django-select2==8.4.3
django-simple-history==3.10.1
django-storages==1.14.6
django-tables2==2.7.5
et_xmlfile==2.0.0
filelock==3.20.0
fontawesomefree==6.6.0
freetype-py==2.5.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.2.0
html5lib==1.1
huggingface-hub==0.36.0
icalendar==6.3.1
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
lunardate==0.2.2
lxml==6.0.2
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
narwhals==2.12.0
networkx==3.6
numpy==2.3.5
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
phonenumbers==9.0.17
pillow==11.3.0
premailer==3.0.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psycopg==3.2.12
pyarrow==21.0.0
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pyluach==2.3.0
PyMeeus==0.5.12
pypdf==6.2.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python-fsutil==0.15.0
python-stdnum==2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
rlPyCairo==0.4.0
rpds-py==0.29.0
s3transfer==0.14.0
safetensors==0.7.0
segno==1.6.6
sentry-sdk==2.26.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8
sqlparse==0.5.3
streamlit==1.51.0
svglib==1.6.0
sympy==1.14.0
tablib==3.9.0
tenacity==9.1.2
timm==1.0.22
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.14
webencodings==0.5.1
workalendar==17.0.0
xhtml2pdf==0.2.17
